Add Authorize using PaymentIntent->capture_method = 'manual'
