+++ src/Action/Api/Resource/CapturePaymentIntentAction.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FluxSE\PayumStripe\Action\Api\Resource;
+
+use FluxSE\PayumStripe\Request\Api\Resource\CapturePaymentIntent;
+use FluxSE\PayumStripe\Request\Api\Resource\RetrieveInterface;
+use Stripe\ApiResource;
+use Stripe\PaymentIntent;
+
+class CapturePaymentIntentAction extends AbstractRetrieveAction
+{
+    protected $apiResourceClass = PaymentIntent::class;
+
+    public function supportAlso(RetrieveInterface $request): bool
+    {
+        return $request instanceof CapturePaymentIntent;
+    }
+
+    public function retrieveApiResource(RetrieveInterface $request): ApiResource
+    {
+        /** @var PaymentIntent $paymentIntent */
+        $paymentIntent = parent::retrieveApiResource($request);
+
+        $paymentIntent->capture();
+
+        return $paymentIntent;
+    }
+}
