--- src/Action/JsConvertPaymentAction.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace FluxSE\PayumStripe\Action;
-
-use Payum\Core\Action\ActionInterface;
-use Payum\Core\Bridge\Spl\ArrayObject;
-use Payum\Core\Exception\RequestNotSupportedException;
-use Payum\Core\GatewayAwareTrait;
-use Payum\Core\Model\PaymentInterface;
-use Payum\Core\Request\Convert;
-
-final class JsConvertPaymentAction implements ActionInterface
-{
-    use GatewayAwareTrait;
-
-    public function execute($request): void
-    {
-        RequestNotSupportedException::assertSupports($this, $request);
-
-        /** @var PaymentInterface $payment */
-        $payment = $request->getSource();
-
-        $details = ArrayObject::ensureArrayObject($payment->getDetails());
-
-        $details->offsetSet('amount', $payment->getTotalAmount());
-        $details->offsetSet('currency', $payment->getCurrencyCode());
-
-        $request->setResult($details);
-    }
-
-    public function supports($request): bool
-    {
-        return
-            $request instanceof Convert &&
-            $request->getSource() instanceof PaymentInterface &&
-            'array' === $request->getTo()
-        ;
-    }
-}
