+++ src/Action/StripeJs/Api/RenderStripeJsAction.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace FluxSE\PayumStripe\Action\Api;
+namespace FluxSE\PayumStripe\Action\StripeJs\Api;
 
-use FluxSE\PayumStripe\Request\Api\Pay;
+use FluxSE\PayumStripe\Action\Api\StripeApiAwareTrait;
+use FluxSE\PayumStripe\Request\StripeJs\Api\RenderStripeJs;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\Exception\RequestNotSupportedException;
@@ -13,7 +14,7 @@ use Payum\Core\GatewayAwareTrait;
 use Payum\Core\Reply\HttpResponse;
 use Payum\Core\Request\RenderTemplate;
 
-final class PayAction implements ActionInterface, ApiAwareInterface, GatewayAwareInterface
+final class RenderStripeJsAction implements ActionInterface, ApiAwareInterface, GatewayAwareInterface
 {
     use GatewayAwareTrait;
     use StripeApiAwareTrait {
@@ -21,21 +22,27 @@ final class PayAction implements ActionInterface, ApiAwareInterface, GatewayAwar
     }
 
     /** @var string */
-    protected $templateName;
+    private $templateName;
 
-    public function __construct(string $templateName)
+    /** @var string */
+    private $apiResourceClass;
+
+    public function __construct(string $templateName, string $apiResourceClass)
     {
         $this->templateName = $templateName;
+        $this->apiResourceClass = $apiResourceClass;
         $this->__stripeApiAwareTraitConstruct();
     }
 
+    /**
+     * @param RenderStripeJs $request
+     */
     public function execute($request): void
     {
         RequestNotSupportedException::assertSupports($this, $request);
 
-        /* @var $request Pay */
         $renderTemplate = new RenderTemplate($this->templateName, [
-            'model' => $request->getPaymentIntent(),
+            'model' => $request->getApiResource(),
             'publishable_key' => $this->api->getPublishableKey(),
             'action_url' => $request->getActionUrl(),
         ]);
@@ -47,6 +54,20 @@ final class PayAction implements ActionInterface, ApiAwareInterface, GatewayAwar
 
     public function supports($request): bool
     {
-        return $request instanceof Pay;
+        if (false === $request instanceof RenderStripeJs) {
+            return false;
+        }
+
+        return $request->getApiResource() instanceof  $this->apiResourceClass;
+    }
+
+    public function getTemplateName(): string
+    {
+        return $this->templateName;
+    }
+
+    public function getApiResourceClass(): string
+    {
+        return $this->apiResourceClass;
     }
 }
