+++ src/Request/Api/Resource/CapturePaymentIntent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FluxSE\PayumStripe\Request\Api\Resource;
+
+final class CapturePaymentIntent extends AbstractRetrieve
+{
+}
