+++ src/Request/Api/Resource/CreateSetupIntent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FluxSE\PayumStripe\Request\Api\Resource;
+
+final class CreateSetupIntent extends AbstractCreate
+{
+}
