+++ src/Request/StripeCheckoutSession/Api/RedirectToCheckout.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FluxSE\PayumStripe\Request\Api;
+namespace FluxSE\PayumStripe\Request\StripeCheckoutSession\Api;
 
 use Payum\Core\Request\Generic;
