+++ src/StripeCheckoutSessionGatewayFactory.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace FluxSE\PayumStripe;
 
-use FluxSE\PayumStripe\Action\Api\RedirectToCheckoutAction;
-use FluxSE\PayumStripe\Action\CaptureAction;
-use FluxSE\PayumStripe\Action\ConvertPaymentAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\Api\RedirectToCheckoutAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\Api\WebhookEvent\CheckoutSessionCompletedAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\AuthorizeAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\CaptureAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\ConvertPaymentAction;
 use Payum\Core\Bridge\Spl\ArrayObject;
 
 final class StripeCheckoutSessionGatewayFactory extends AbstractStripeGatewayFactory
@@ -16,15 +18,19 @@ final class StripeCheckoutSessionGatewayFactory extends AbstractStripeGatewayFac
         parent::populateConfig($config);
 
         $config->defaults([
-            // Factory
+            // Factories
             'payum.factory_name' => 'stripe_checkout_session',
             'payum.factory_title' => 'Stripe Checkout Session',
 
             // Templates
             'payum.template.redirect_to_checkout' => '@FluxSEPayumStripe/Action/redirectToCheckout.html.twig',
 
-            // Action
+            // Webhook event resolver
+            'payum.action.checkout_session_completed' => new CheckoutSessionCompletedAction(),
+
+            // Actions
             'payum.action.capture' => new CaptureAction(),
+            'payum.action.authorize' => new AuthorizeAction(),
             'payum.action.convert_payment' => new ConvertPaymentAction(),
             'payum.action.redirect_to_checkout' => function (ArrayObject $config) {
                 return new RedirectToCheckoutAction($config['payum.template.redirect_to_checkout']);
