+++ src/StripeJsGatewayFactory.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace FluxSE\PayumStripe;
 
-use FluxSE\PayumStripe\Action\Api\PayAction;
-use FluxSE\PayumStripe\Action\JsCaptureAction;
-use FluxSE\PayumStripe\Action\JsConvertPaymentAction;
+use FluxSE\PayumStripe\Action\StripeJs\Api\RenderStripeJsAction;
+use FluxSE\PayumStripe\Action\StripeJs\AuthorizeAction;
+use FluxSE\PayumStripe\Action\StripeJs\CaptureAction;
+use FluxSE\PayumStripe\Action\StripeJs\ConvertPaymentAction;
 use Payum\Core\Bridge\Spl\ArrayObject;
+use Stripe\PaymentIntent;
 
 final class StripeJsGatewayFactory extends AbstractStripeGatewayFactory
 {
@@ -16,18 +18,22 @@ final class StripeJsGatewayFactory extends AbstractStripeGatewayFactory
         parent::populateConfig($config);
 
         $config->defaults([
-            // Factory
+            // Factories
             'payum.factory_name' => 'stripe_js',
             'payum.factory_title' => 'Stripe JS',
 
             // Templates
-            'payum.template.pay' => '@FluxSEPayumStripe/Action/pay.html.twig',
+            'payum.template.render_stripe_js.payment_intent' => '@FluxSEPayumStripe/Action/stripeJsPaymentIntent.html.twig',
 
             // Actions
-            'payum.action.capture' => new JsCaptureAction(),
-            'payum.action.convert_payment' => new JsConvertPaymentAction(),
-            'payum.action.pay' => function (ArrayObject $config) {
-                return new PayAction($config['payum.template.pay']);
+            'payum.action.capture' => new CaptureAction(),
+            'payum.action.authorize' => new AuthorizeAction(),
+            'payum.action.convert_payment' => new ConvertPaymentAction(),
+            'payum.action.render_stripe_js.payment_intent' => function (ArrayObject $config) {
+                return new RenderStripeJsAction(
+                    $config['payum.template.render_stripe_js.payment_intent'],
+                    PaymentIntent::class
+                );
             },
         ]);
     }
