+++ tests/Action/StripeCheckoutSession/ConvertPaymentActionTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Tests\FluxSE\PayumStripe\Action;
+namespace Tests\FluxSE\PayumStripe\Action\StripeCheckoutSession;
 
-use FluxSE\PayumStripe\Action\JsConvertPaymentAction;
+use FluxSE\PayumStripe\Action\StripeCheckoutSession\ConvertPaymentAction;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\GatewayInterface;
@@ -10,11 +10,11 @@ use Payum\Core\Model\Payment;
 use Payum\Core\Request\Convert;
 use PHPUnit\Framework\TestCase;
 
-final class JsConvertPaymentActionTest extends TestCase
+final class ConvertPaymentActionTest extends TestCase
 {
     public function testShouldImplements()
     {
-        $action = new JsConvertPaymentAction();
+        $action = new ConvertPaymentAction();
 
         $this->assertInstanceOf(ActionInterface::class, $action);
         $this->assertNotInstanceOf(GatewayInterface::class, $action);
@@ -24,12 +24,14 @@ final class JsConvertPaymentActionTest extends TestCase
     public function testShouldCorrectlyConvertPaymentToDetailsAndSetItBack()
     {
         $payment = new Payment();
+        $payment->setClientEmail('test@domain.tld');
+        $payment->setDescription('the description');
         $payment->setTotalAmount(123);
         $payment->setCurrencyCode('USD');
 
         $request = new Convert($payment, 'array');
 
-        $action = new JsConvertPaymentAction();
+        $action = new ConvertPaymentAction();
 
         $supports = $action->supports($request);
         $this->assertTrue($supports);
@@ -39,25 +41,37 @@ final class JsConvertPaymentActionTest extends TestCase
         $details = $request->getResult();
 
         $this->assertNotEmpty($details);
-        $this->assertArrayHasKey('amount', $details);
-        $this->assertArrayHasKey('currency', $details);
+        $this->assertArrayHasKey('customer_email', $details);
+        $this->assertArrayHasKey('line_items', $details);
+        $this->assertIsArray($details['line_items']);
+        $this->assertIsArray($details['line_items'][0]);
+        $this->assertArrayHasKey('name', $details['line_items'][0]);
+        $this->assertArrayHasKey('amount', $details['line_items'][0]);
+        $this->assertArrayHasKey('currency', $details['line_items'][0]);
+        $this->assertArrayHasKey('quantity', $details['line_items'][0]);
+        $this->assertArrayHasKey('payment_method_types', $details);
 
-        $this->assertEquals(123, $details['amount']);
-        $this->assertEquals('USD', $details['currency']);
+        $this->assertEquals('test@domain.tld', $details['customer_email']);
+        $this->assertEquals('the description', $details['line_items'][0]['name']);
+        $this->assertEquals(123, $details['line_items'][0]['amount']);
+        $this->assertEquals('USD', $details['line_items'][0]['currency']);
+        $this->assertEquals(1, $details['line_items'][0]['quantity']);
     }
 
     public function testShouldNotOverwriteAlreadySetExtraDetails()
     {
         $payment = new Payment();
+        $payment->setClientEmail('test@domain.tld');
         $payment->setCurrencyCode('USD');
         $payment->setTotalAmount(123);
+        $payment->setDescription('the description');
         $payment->setDetails([
             'foo' => 'fooVal',
         ]);
 
         $request = new Convert($payment, 'array');
 
-        $action = new JsConvertPaymentAction();
+        $action = new ConvertPaymentAction();
 
         $supports = $action->supports($request);
         $this->assertTrue($supports);
