+++ tests/Action/StripeJs/Api/RenderStripeJsActionTest.php
@@ -1,52 +1,84 @@
 <?php
 
-namespace Tests\FluxSE\PayumStripe\Action\Api;
+namespace Tests\FluxSE\PayumStripe\Action\StripeJs\Api;
 
-use ArrayObject;
-use FluxSE\PayumStripe\Action\Api\RedirectToCheckoutAction;
+use FluxSE\PayumStripe\Action\StripeJs\Api\RenderStripeJsAction;
 use FluxSE\PayumStripe\Api\KeysInterface;
-use FluxSE\PayumStripe\Request\Api\RedirectToCheckout;
+use FluxSE\PayumStripe\Request\Api\ConstructEvent;
+use FluxSE\PayumStripe\Request\StripeJs\Api\RenderStripeJs;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\GatewayAwareInterface;
 use Payum\Core\Reply\HttpResponse;
 use Payum\Core\Request\RenderTemplate;
 use PHPUnit\Framework\TestCase;
+use Stripe\ApiResource;
+use Stripe\Checkout\Session;
+use Stripe\PaymentIntent;
+use Stripe\SetupIntent;
+use Tests\FluxSE\PayumStripe\Action\Api\ApiAwareActionTestTrait;
 use Tests\FluxSE\PayumStripe\Action\GatewayAwareTestTrait;
 
-final class RedirectToCheckoutActionTest extends TestCase
+final class RenderStripeJsActionTest extends TestCase
 {
     use ApiAwareActionTestTrait;
     use GatewayAwareTestTrait;
 
+    public function apiRessourcesList(): array
+    {
+        return [
+            [PaymentIntent::class],
+            [SetupIntent::class],
+        ];
+    }
+
     public function testShouldImplements()
     {
-        $action = new RedirectToCheckoutAction('aTemplateName');
+        $action = new RenderStripeJsAction('aTemplateName', ApiResource::class);
 
         $this->assertInstanceOf(ActionInterface::class, $action);
         $this->assertInstanceOf(ApiAwareInterface::class, $action);
         $this->assertInstanceOf(GatewayAwareInterface::class, $action);
     }
 
-    public function testShouldNotSupportObtainTokenRequestWithNotArrayAccessModel()
+    /**
+     * @dataProvider apiRessourcesList
+     */
+    public function testShouldSupportOnlyATypedRequestAndAnApiRessourceClass(string $apiResourceClass)
     {
-        $model = [];
-        $action = new RedirectToCheckoutAction('aTemplateName');
+        $model = new $apiResourceClass();
+        $action = new RenderStripeJsAction('aTemplateName', $apiResourceClass);
+
+        $this->assertTrue($action->supports(new RenderStripeJs($model, '')));
+        $this->assertFalse($action->supports(new RenderStripeJs(new Session(), '')));
+        $this->assertFalse($action->supports(new ConstructEvent('', '', '')));
+    }
+
+    /**
+     * @dataProvider apiRessourcesList
+     */
+    public function testShouldNotSupportObtainTokenRequestWithNotArrayAccessModel(string $apiResourceClass)
+    {
+        $model = new $apiResourceClass();
+        $actionUrl = '';
+        $action = new RenderStripeJsAction('aTemplateName', $apiResourceClass);
 
         $gatewayMock = $this->createGatewayMock();
         $gatewayMock
             ->expects($this->once())
             ->method('execute')
             ->with($this->isInstanceOf(RenderTemplate::class))
-            ->will($this->returnCallback(function (RenderTemplate $request) use ($model) {
+            ->will($this->returnCallback(function (RenderTemplate $request) use ($model, $actionUrl) {
                 $parameters = $request->getParameters();
                 $this->assertEquals('aTemplateName', $request->getTemplateName());
                 $this->assertIsArray($parameters);
                 $this->assertArrayHasKey('model', $parameters);
                 $this->assertArrayHasKey('publishable_key', $parameters);
+                $this->assertArrayHasKey('action_url', $parameters);
                 $this->assertEquals([
-                    'model' => new ArrayObject($model),
+                    'model' => $model,
                     'publishable_key' => '',
+                    'action_url' => $actionUrl,
                 ], $parameters);
                 $request->setResult('');
             }));
@@ -62,7 +94,7 @@ final class RedirectToCheckoutActionTest extends TestCase
         $action->setGateway($gatewayMock);
         $action->setApi($apiMock);
 
-        $request = new RedirectToCheckout($model);
+        $request = new RenderStripeJs($model, $actionUrl);
 
         $supports = $action->supports($request);
         $this->assertTrue($supports);
