+++ tests/Action/StripeJs/ConvertPaymentActionTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Tests\FluxSE\PayumStripe\Action;
+namespace Tests\FluxSE\PayumStripe\Action\StripeJs;
 
-use FluxSE\PayumStripe\Action\ConvertPaymentAction;
+use FluxSE\PayumStripe\Action\StripeJs\ConvertPaymentAction;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\GatewayInterface;
@@ -24,8 +24,6 @@ final class ConvertPaymentActionTest extends TestCase
     public function testShouldCorrectlyConvertPaymentToDetailsAndSetItBack()
     {
         $payment = new Payment();
-        $payment->setClientEmail('test@domain.tld');
-        $payment->setDescription('the description');
         $payment->setTotalAmount(123);
         $payment->setCurrencyCode('USD');
 
@@ -41,30 +39,18 @@ final class ConvertPaymentActionTest extends TestCase
         $details = $request->getResult();
 
         $this->assertNotEmpty($details);
-        $this->assertArrayHasKey('customer_email', $details);
-        $this->assertArrayHasKey('line_items', $details);
-        $this->assertIsArray($details['line_items']);
-        $this->assertIsArray($details['line_items'][0]);
-        $this->assertArrayHasKey('name', $details['line_items'][0]);
-        $this->assertArrayHasKey('amount', $details['line_items'][0]);
-        $this->assertArrayHasKey('currency', $details['line_items'][0]);
-        $this->assertArrayHasKey('quantity', $details['line_items'][0]);
-        $this->assertArrayHasKey('payment_method_types', $details);
+        $this->assertArrayHasKey('amount', $details);
+        $this->assertArrayHasKey('currency', $details);
 
-        $this->assertEquals('test@domain.tld', $details['customer_email']);
-        $this->assertEquals('the description', $details['line_items'][0]['name']);
-        $this->assertEquals(123, $details['line_items'][0]['amount']);
-        $this->assertEquals('USD', $details['line_items'][0]['currency']);
-        $this->assertEquals(1, $details['line_items'][0]['quantity']);
+        $this->assertEquals(123, $details['amount']);
+        $this->assertEquals('USD', $details['currency']);
     }
 
     public function testShouldNotOverwriteAlreadySetExtraDetails()
     {
         $payment = new Payment();
-        $payment->setClientEmail('test@domain.tld');
         $payment->setCurrencyCode('USD');
         $payment->setTotalAmount(123);
-        $payment->setDescription('the description');
         $payment->setDetails([
             'foo' => 'fooVal',
         ]);
